changed vars to size_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,9 +7,9 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned int count;
+	size_t count;
 	const list_t *current;
-	unsigned int i;
+	size_t i;
 
 	count = 1;
 	current = h;
